requirements.txt
anyio==4.9.0
asgiref==3.8.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.0
django-filter==25.1
django-model-utils==5.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
firebase-admin==6.9.0
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
mailjet-rest==1.5.1
msgpack==1.1.0
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg-binary==3.2.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.2.0
urllib3==2.4.0
whitenoise==6.9.0
mailjet-rest==1.3.4
stripe==11.1.0
